pyproject.toml
[project]
name = "pmi-agent"
version = "0.1.0"
description = "NTA Pre-Market Intelligence Agent — Phase 1 reference build"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Eoin Lane" },
    { name = "Cathal Bellew" },
]
dependencies = [
    "anthropic>=0.40.0",
    "openpyxl>=3.1",
    "pandas>=2.2",
    "pydantic>=2.7",
    "python-docx>=1.1",
    "python-dotenv>=1.0",
    "streamlit>=1.30",
]

[dependency-groups]
dev = [
    "pyright>=1.1",
    "pytest>=8.0",
    "ruff>=0.5",
]

[project.scripts]
pmi-spec-analyse = "pmi_agent.cli.spec_analyse:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pmi_agent"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
